convert coords into float

+++ unstructured/partition/pdf_image/inference_utils.py
@@ -93,10 +93,10 @@ def merge_text_regions(regions: TextRegions) -> TextRegion:
     if not regions:
         raise ValueError("The text regions to be merged must be provided.")
 
-    min_x1 = regions.x1.min()
-    min_y1 = regions.y1.min()
-    max_x2 = regions.x2.max()
-    max_y2 = regions.y2.max()
+    min_x1 = regions.x1.min().astype(float)
+    min_y1 = regions.y1.min().astype(float)
+    max_x2 = regions.x2.max().astype(float)
+    max_y2 = regions.y2.max().astype(float)
 
     merged_text = " ".join(regions.texts)
     source = regions.source
